--- Data/sec_data/missing_ciks_2010.csv
@@ -1,5 +1,0 @@
-ticker
-BF.B
-CBS
-LSI
-VIAb
