--- Data/sec_data/missing_ciks_2025.csv
@@ -1,2 +1,0 @@
-ticker
-BF.B
