--- Data/sp500_tickers_2010.csv
@@ -1,500 +1,0 @@
-Ticker
-MMM
-ABT
-ANF
-ADBE
-AMD
-AES
-AET
-ACS
-AFL
-A
-APD
-AKAM
-AKS
-AA
-AYE
-ATI
-AGN
-ALL
-ALTR
-MO
-AMZN
-AEE
-AEP
-AXP
-AIG
-AMT
-AMP
-ABC
-AMGN
-APC
-ADI
-AON
-APA
-AIV
-APOL
-AAPL
-AMAT
-ADM
-ASH
-AIZ
-T
-ADSK
-ADP
-AN
-AZO
-AVB
-AVY
-AVP
-BHI
-BLL
-BAC
-BK
-BCR
-BAX
-BBT
-BDX
-BBBY
-BMS
-BRK.B
-BBY
-BIG
-BIIB
-BJS
-BDK
-HRB
-BMC
-BA
-BXP
-BSX
-BMY
-BRCM
-BF.B
-CHRW
-CA
-COG
-CAM
-CPB
-COF
-CAH
-CFN
-CCL
-CAT
-CBG
-CBS
-CELG
-CNP
-CTL
-CEPH
-CF
-SCHW
-CHK
-CVX
-CB
-CI
-CINF
-CTAS
-CSCO
-C
-CTXS
-CLF
-CLX
-CME
-CMS
-COH
-KO
-CCE
-CTSH
-CL
-CMCSA
-CMA
-CSC
-CPWR
-CAG
-COP
-CNX
-ED
-STZ
-CEG
-CBE
-GLW
-COST
-CVH
-CSX
-CMI
-CVS
-DHI
-DHR
-DRI
-DVA
-DF
-DE
-DELL
-DNR
-XRAY
-DVN
-DV
-DO
-DTV
-DFS
-D
-RRD
-DOV
-DOW
-DPS
-DTE
-DD
-DUK
-DNB
-ETFC
-EMN
-EK
-ETN
-EBAY
-ECL
-EIX
-EP
-ERTS
-EMC
-EMR
-ETR
-EOG
-EQT
-EFX
-EQR
-EL
-EXC
-EXPE
-EXPD
-ESRX
-XOM
-FDO
-FAST
-FII
-FDX
-FIS
-FITB
-FHN
-FSLR
-FE
-FISV
-FLIR
-FLS
-FLR
-FMC
-FTI
-F
-FRX
-FO
-FPL
-BEN
-FCX
-FTR
-GME
-GCI
-GPS
-GD
-GE
-GIS
-GPC
-GNW
-GENZ
-GILD
-GS
-GR
-GT
-GOOG
-GWW
-HAL
-HOG
-HAR
-HRS
-HIG
-HAS
-HCP
-HCN
-HNZ
-HES
-HPQ
-HD
-HON
-HRL
-HSP
-HST
-HCBK
-HUM
-HBAN
-ITW
-RX
-TEG
-INTC
-ICE
-IBM
-IFF
-IGT
-IP
-IPG
-INTU
-ISRG
-IVZ
-IRM
-ITT
-JBL
-JEC
-JNS
-JDSU
-JNJ
-JCI
-JPM
-JNPR
-K
-KEY
-KMB
-KIM
-KG
-KLAC
-KSS
-KFT
-KR
-LLL
-LH
-LM
-LEG
-LEN
-LUK
-LXK
-LIFE
-LLY
-LTD
-LNC
-LLTC
-LMT
-L
-LO
-LOW
-LSI
-MTB
-M
-MRO
-MAR
-MMC
-MI
-MAS
-MEE
-MA
-MAT
-MFE
-MKC
-MCD
-MHP
-MCK
-MJN
-MWV
-MHS
-MDT
-WFR
-MRK
-MDP
-MET
-PCS
-MCHP
-MU
-MSFT
-MIL
-MOLX
-TAP
-MON
-MWW
-MCO
-MS
-MOT
-MXB
-MUR
-MYL
-NBR
-NDAQ
-NOV
-NSM
-NTAP
-NYT
-NWL
-NEM
-NWSA
-GAS
-NKE
-NI
-NBL
-JWN
-NSC
-NTRS
-NOC
-NU
-NOVL
-NVLS
-NUE
-NVDA
-NYX
-ORLY
-OXY
-ODP
-OMC
-ORCL
-OI
-PCAR
-PTV
-PLL
-PH
-PDCO
-PAYX
-BTU
-JCP
-PBCT
-POM
-PBG
-PEP
-PKI
-PFE
-PCG
-PM
-PNW
-PXD
-PBI
-PCL
-PNC
-RL
-PPG
-PPL
-PX
-PCP
-PFG
-PG
-PGN
-PGR
-PLD
-PRU
-PEG
-PSA
-PHM
-QLGC
-PWR
-QCOM
-DGX
-STR
-Q
-RSH
-RRC
-RTN
-RHT
-RF
-RSG
-RAI
-RHI
-ROK
-COL
-ROP
-ROST
-RDC
-R
-SWY
-SAI
-CRM
-SNDK
-SLE
-SCG
-SLB
-SNI
-SEE
-SHLD
-SRE
-SHW
-SIAL
-SPG
-SLM
-SII
-SJM
-SNA
-SO
-LUV
-SWN
-SE
-S
-STJ
-SWK
-SPLS
-SBUX
-HOT
-STT
-SRCL
-SYK
-SUN
-STI
-SVU
-SYMC
-SYY
-TROW
-TGT
-TE
-TLAB
-THC
-TDC
-TER
-TSO
-TXN
-TXT
-HSY
-TRV
-TMO
-TIF
-TWX
-TWC
-TIE
-TJX
-TMK
-TSS
-TSN
-USB
-UNP
-UNH
-UPS
-X
-UTX
-UNM
-VFC
-VLO
-VAR
-VTR
-VRSN
-VZ
-VIAb
-V
-VNO
-VMC
-WMT
-WAG
-DIS
-WPO
-WM
-WAT
-WPI
-WLP
-WFC
-WDC
-WU
-WY
-WHR
-WFMI
-WMB
-WIN
-WEC
-WYN
-WYNN
-XEL
-XRX
-XLNX
-XL
-XTO
-YHOO
-YUM
-ZMH
-ZION
