+++ data/intermidiate/missing_ciks_2010.csv
@@ -1,0 +1,5 @@
+ticker
+BF.B
+CBS
+LSI
+VIAb
