+++ data/intermidiate/missing_ciks_2019.csv
@@ -1,0 +1,17 @@
+ticker
+ANTM
+ARNC
+BHGE
+BF.B
+CBS
+DISCK
+DWDP
+FB
+FRC
+FBHS
+HFC
+INFO
+NLSN
+TWTR
+VIAB
+WLTW
