+++ data/tickers/sp500_tickers_2019.csv
@@ -1,0 +1,506 @@
+Ticker
+MMM
+ABT
+ABBV
+ABMD
+ACN
+ATVI
+ADBE
+AMD
+AAP
+AES
+AMG
+AFL
+A
+APD
+AKAM
+ALK
+ALB
+ARE
+ALXN
+ALGN
+ALLE
+AGN
+ADS
+LNT
+ALL
+GOOGL
+GOOG
+MO
+AMZN
+AEE
+AAL
+AEP
+AXP
+AIG
+AMT
+AWK
+AMP
+ABC
+AME
+AMGN
+APH
+APC
+ADI
+ANSS
+ANTM
+AON
+AOS
+APA
+AIV
+AAPL
+AMAT
+APTV
+ADM
+ARNC
+ANET
+AJG
+AIZ
+T
+ADSK
+ADP
+AZO
+AVB
+AVY
+BHGE
+BLL
+BAC
+BK
+BAX
+BBT
+BDX
+BRK.B
+BBY
+BIIB
+BLK
+HRB
+BA
+BKNG
+BWA
+BXP
+BSX
+BHF
+BMY
+AVGO
+BR
+BF.B
+CHRW
+COG
+CDNS
+CPB
+COF
+CPRI
+CAH
+KMX
+CCL
+CAT
+CBOE
+CBRE
+CBS
+CE
+CELG
+CNC
+CNP
+CTL
+CERN
+CF
+SCHW
+CHTR
+CVX
+CMG
+CB
+CHD
+CI
+XEC
+CINF
+CTAS
+CSCO
+C
+CFG
+CTXS
+CLX
+CME
+CMS
+KO
+CTSH
+CL
+CMCSA
+CMA
+CAG
+CXO
+COP
+ED
+STZ
+COO
+CPRT
+GLW
+COST
+COTY
+CCI
+CSX
+CMI
+CVS
+DHI
+DHR
+DRI
+DVA
+DE
+DAL
+XRAY
+DVN
+FANG
+DLR
+DFS
+DISCA
+DISCK
+DISH
+DG
+DLTR
+D
+DOV
+DWDP
+DTE
+DRE
+DUK
+DXC
+ETFC
+EMN
+ETN
+EBAY
+ECL
+EIX
+EW
+EA
+EMR
+ETR
+EOG
+EFX
+EQIX
+EQR
+ESS
+EL
+EVRG
+ES
+RE
+EXC
+EXPE
+EXPD
+EXR
+XOM
+FFIV
+FB
+FAST
+FRT
+FDX
+FIS
+FITB
+FE
+FRC
+FISV
+FLT
+FLIR
+FLS
+FLR
+FMC
+FL
+F
+FTNT
+FTV
+FBHS
+BEN
+FCX
+GPS
+GRMN
+IT
+GD
+GE
+GIS
+GM
+GPC
+GILD
+GPN
+GS
+GT
+GWW
+HAL
+HBI
+HOG
+HRS
+HIG
+HAS
+HCA
+HCP
+HP
+HSIC
+HSY
+HES
+HPE
+HLT
+HFC
+HOLX
+HD
+HON
+HRL
+HST
+HPQ
+HUM
+HBAN
+HII
+IDXX
+INFO
+ITW
+ILMN
+IR
+INTC
+ICE
+IBM
+INCY
+IP
+IPG
+IFF
+INTU
+ISRG
+IVZ
+IPGP
+IQV
+IRM
+JKHY
+JEC
+JBHT
+JEF
+SJM
+JNJ
+JCI
+JPM
+JNPR
+KSU
+K
+KEY
+KEYS
+KMB
+KIM
+KMI
+KLAC
+KSS
+KHC
+KR
+LB
+LLL
+LH
+LRCX
+LW
+LEG
+LEN
+LLY
+LNC
+LIN
+LKQ
+LMT
+L
+LOW
+LYB
+MTB
+MAC
+M
+MRO
+MPC
+MAR
+MMC
+MLM
+MAS
+MA
+MAT
+MKC
+MXIM
+MCD
+MCK
+MDT
+MRK
+MET
+MTD
+MGM
+MCHP
+MU
+MSFT
+MAA
+MHK
+TAP
+MDLZ
+MNST
+MCO
+MS
+MOS
+MSI
+MSCI
+MYL
+NDAQ
+NOV
+NKTR
+NTAP
+NFLX
+NWL
+NFX
+NEM
+NWSA
+NWS
+NEE
+NLSN
+NKE
+NI
+NBL
+JWN
+NSC
+NTRS
+NOC
+NCLH
+NRG
+NUE
+NVDA
+ORLY
+OXY
+OMC
+OKE
+ORCL
+PCAR
+PKG
+PH
+PAYX
+PYPL
+PNR
+PBCT
+PEP
+PKI
+PRGO
+PFE
+PCG
+PM
+PSX
+PNW
+PXD
+PNC
+RL
+PPG
+PPL
+PFG
+PG
+PGR
+PLD
+PRU
+PEG
+PSA
+PHM
+PVH
+QRVO
+PWR
+QCOM
+DGX
+RJF
+RTN
+O
+RHT
+REG
+REGN
+RF
+RSG
+RMD
+RHI
+ROK
+ROL
+ROP
+ROST
+RCL
+CRM
+SBAC
+SLB
+STX
+SEE
+SRE
+SHW
+SPG
+SWKS
+SLG
+SNA
+SO
+LUV
+SPGI
+SWK
+SBUX
+STT
+SYK
+STI
+SIVB
+SYMC
+SYF
+SNPS
+SYY
+TROW
+TTWO
+TPR
+TGT
+TEL
+FTI
+TXN
+TXT
+TMO
+TIF
+TWTR
+TJX
+TMK
+TSS
+TSCO
+TDG
+TRV
+TRIP
+FOXA
+FOX
+TSN
+UDR
+ULTA
+USB
+UAA
+UA
+UNP
+UAL
+UNH
+UPS
+URI
+UTX
+UHS
+UNM
+VFC
+VLO
+VAR
+VTR
+VRSN
+VRSK
+VZ
+VRTX
+VIAB
+V
+VNO
+VMC
+WMT
+WBA
+DIS
+WM
+WAT
+WEC
+WCG
+WFC
+WELL
+WDC
+WU
+WRK
+WY
+WHR
+WMB
+WLTW
+WYNN
+XEL
+XRX
+XLNX
+XYL
+YUM
+ZBH
+ZION
+ZTS
